Make experimental-discord the canonical discord plugin
The naming of "experimental-discord" was incorrect since it's the only Discord plugin in actual use.
This updates the name to "discord" while renaming the current "discord" plugin to "wip-discord" since
it is not yet feature complete and might not be developed further.

Test Plan: Ensure CI tests pass
