+++ src/plugins/discord/config.js
