+++ src/plugins/discord/declaration.js
