+++ src/plugins/discord/fetcher.js
@@ -1,79 +1,126 @@
 // @flow
 
+import fetch from "isomorphic-fetch";
 import * as Model from "./models";
-import {type Snowflake} from "./models";
 
-type FetchEndpoint = (endpoint: string) => Promise<any>;
+export interface DiscordApi {
+  guilds(): Promise<$ReadOnlyArray<Model.Guild>>;
+  emojis(guild: Model.Snowflake): Promise<$ReadOnlyArray<Model.Emoji>>;
+  channels(guild: Model.Snowflake): Promise<$ReadOnlyArray<Model.Channel>>;
+  roles(guild: Model.Snowflake): Promise<$ReadOnlyArray<Model.Role>>;
+  members(guild: Model.Snowflake): Promise<$ReadOnlyArray<Model.GuildMember>>;
+  messages(
+    channel: Model.Snowflake,
+    after: Model.Snowflake,
+    limit: number
+  ): Promise<$ReadOnlyArray<Model.Message>>;
+  reactions(
+    channel: Model.Snowflake,
+    message: Model.Snowflake,
+    emoji: Model.Emoji
+  ): Promise<$ReadOnlyArray<Model.Reaction>>;
+}
 
-/**
- * Provide the Guild ID to fetch against, and the 'limit'
- * parameter when fetching GuildMembers, Messages, and Reactions.
- */
-export type FetchOptions = {|
-  membersLimit: number,
-  messagesLimit: number,
-  reactionsLimit: number,
+const fetcherDefaults: FetcherOptions = {
+  apiUrl: "https://discordapp.com/api",
+  token: null,
+  fetch,
+};
+
+type FetcherOptions = {|
+  +apiUrl: string,
+  +fetch: typeof fetch,
+  +token: ?Model.BotToken,
 |};
 
-export type PageInfo = {|
-  +hasNextPage: boolean,
-  +endCursor: Snowflake | null,
-|};
+export class Fetcher implements DiscordApi {
+  +_options: FetcherOptions;
 
-export type ResultPage<T> = {|
-  +pageInfo: PageInfo,
-  +results: $ReadOnlyArray<T>,
-|};
-
-/**
- * Fetcher is responsible for:
- * - Returning the correct endpoint to fetch against for Guilds, Channels,
- *   Members, and Reactions.
- * - Formatting the returned results into the correct Typed objects
- * - Returning pagination info in a PageInfo object, containing hasNextPage
- *   and endCursor properties.
- *   The endCursor property is calculated as the Id of the last object
- *   returned in the response results. We are assuming Discord provides
- *   consistent, ordered results.
- *   The hasNextPage property is a boolean calculated as whether the number of
- *   results recieved is equal to the `limit` property provided in the
- *   fetch request.
- *
- *   Note that Discord doesn't support pagination for Channels, so we're
- *   returning an array of Channel objects in the corresponding method.
- *   See: https://discordapp.com/developers/docs/resources/guild#get-guild-channels
- */
-export class DiscordFetcher {
-  +_fetch: FetchEndpoint;
-  +_options: FetchOptions;
-
-  constructor(fetchEndpoint: FetchEndpoint, options: FetchOptions) {
-    this._fetch = fetchEndpoint;
-    this._options = options;
+  constructor(opts?: $Shape<FetcherOptions>) {
+    this._options = {...fetcherDefaults, ...opts};
+    if (!this._options.token) {
+      throw new Error("A BotToken is required");
+    }
   }
 
-  async guild(guildId: Snowflake): Promise<Model.Guild> {
-    const {id, name, permissions} = await this._fetch(`guilds/${guildId}`);
-    return {id: id, name: name, permissions: permissions};
+  _fetch(endpoint: string): Promise<Response> {
+    const {apiUrl, token} = this._options;
+    if (!token) {
+      throw new Error("A BotToken is required");
+    }
+    const requestOptions = {
+      method: "GET",
+      headers: {
+        Accept: "application/json",
+        Authorization: `Bot ${token}`,
+      },
+    };
+    const url = new URL(`${apiUrl}${endpoint}`).href;
+    return this._options.fetch(url, requestOptions);
   }
 
-  async channels(guildId: Snowflake): Promise<$ReadOnlyArray<Model.Channel>> {
-    const response = await this._fetch(`/guilds/${guildId}/channels`);
-    return response.map((x) => ({
+  async _fetchJson(endpoint: string): Promise<any> {
+    const res = await this._fetch(endpoint);
+    failIfMissing(res);
+    failForNotOk(res);
+    return await res.json();
+  }
+
+  async guilds(): Promise<$ReadOnlyArray<Model.Guild>> {
+    const guilds = await this._fetchJson("/users/@me/guilds");
+    return guilds.map((x) => ({
+      id: x.id,
+      name: x.name,
+      permissions: x.permissions,
+    }));
+  }
+
+  async emojis(guild: Model.Snowflake): Promise<$ReadOnlyArray<Model.Emoji>> {
+    const emojis = await this._fetchJson(`/guilds/${guild}/emojis`);
+    return emojis.map((x) => ({
+      id: x.id,
+      name: x.name,
+    }));
+  }
+
+  async channels(
+    guild: Model.Snowflake
+  ): Promise<$ReadOnlyArray<Model.Channel>> {
+    const channels = await this._fetchJson(`/guilds/${guild}/channels`);
+    return channels.map((x) => ({
       id: x.id,
       name: x.name,
       type: Model.channelTypeFromId(x.type),
     }));
   }
 
+  async roles(guild: Model.Snowflake): Promise<$ReadOnlyArray<Model.Role>> {
+    const roles = await this._fetchJson(`/guilds/${guild}/roles`);
+    return roles.map((x) => ({
+      id: x.id,
+      name: x.name,
+    }));
+  }
+
   async members(
-    guildId: Snowflake,
-    after: Snowflake
-  ): Promise<ResultPage<Model.GuildMember>> {
-    const {membersLimit} = this._options;
-    const endpoint = `/guilds/${guildId}/members?after=${after}&limit=${membersLimit}`;
-    const response = await this._fetch(endpoint);
-    const results = response.map((x) => ({
+    guild: Model.Snowflake
+  ): Promise<$ReadOnlyArray<Model.GuildMember>> {
+    const limit = 1000;
+    let doneLoading = false;
+    let allMembers = [];
+    let after = "0";
+    while (!doneLoading) {
+      const newMembers: Array<Model.GuildMember> = await this._fetchJson(
+        `/guilds/${guild}/members?after=${after}&limit=${limit}`
+      );
+      if (newMembers.length < limit) {
+        doneLoading = true;
+      } else {
+        after = newMembers[newMembers.length - 1].user.id;
+      }
+      allMembers = [...allMembers, ...newMembers];
+    }
+    return allMembers.map((x) => ({
       user: {
         id: x.user.id,
         username: x.user.username,
@@ -81,57 +128,75 @@ export class DiscordFetcher {
         bot: x.user.bot || x.user.system || false,
       },
       nick: x.nick || null,
+      roles: x.roles,
     }));
-    const hasNextPage = results.length === membersLimit;
-    const endCursor =
-      response.length > 0 ? response[response.length - 1].user.id : null;
-    const pageInfo = {hasNextPage, endCursor};
-    return {results, pageInfo};
   }
 
   async messages(
-    channel: Snowflake,
-    after: Snowflake
-  ): Promise<ResultPage<Model.Message>> {
-    const {messagesLimit} = this._options;
-    const endpoint = `/channels/${channel}/messages?after=${after}&limit=${messagesLimit}`;
-    const response = await this._fetch(endpoint);
-    const results = response.map((x) => ({
+    channel: Model.Snowflake,
+    after: Model.Snowflake,
+    limit: number
+  ): Promise<$ReadOnlyArray<Model.Message>> {
+    const messages = await this._fetchJson(
+      `/channels/${channel}/messages?after=${after}&limit=${limit}`
+    );
+    return messages.map((x) => ({
       id: x.id,
       channelId: channel,
       authorId: x.author.id,
       timestampMs: Date.parse(x.timestamp),
       content: x.content,
       reactionEmoji: (x.reactions || []).map((r) => r.emoji),
-      nonUserAuthor: Model.isAuthoredByNonUser(x),
+      nonUserAuthor: x.webhook_id != null || false,
       mentions: (x.mentions || []).map((user) => user.id),
     }));
-    const hasNextPage = results.length === messagesLimit;
-    const endCursor = response.length > 0 ? response[0].id : null;
-    const pageInfo = {hasNextPage, endCursor};
-    return {results, pageInfo};
   }
 
   async reactions(
-    channel: Snowflake,
-    message: Snowflake,
-    emoji: Model.Emoji,
-    after: Snowflake
-  ): Promise<ResultPage<Model.Reaction>> {
-    const {reactionsLimit} = this._options;
+    channel: Model.Snowflake,
+    message: Model.Snowflake,
+    emoji: Model.Emoji
+  ): Promise<$ReadOnlyArray<Model.Reaction>> {
+    let doneLoading = false;
+    let allReactingUsers: Array<Model.User> = [];
+    let after = "0";
+    const limit = 100;
     const emojiRef = Model.emojiToRef(emoji);
-    const endpoint = `/channels/${channel}/messages/${message}/reactions/${emojiRef}?after=${after}&limit=${reactionsLimit}`;
-    const response = await this._fetch(endpoint);
-    const results = response.map((x) => ({
-      emoji: emoji,
+    while (!doneLoading) {
+      const newReactingUsers: Array<Model.User> = await this._fetchJson(
+        `/channels/${channel}/messages/${message}/reactions/${emojiRef}?after=${after}&limit=${limit}`
+      );
+      if (newReactingUsers.length < limit) {
+        doneLoading = true;
+      } else {
+        after = newReactingUsers[newReactingUsers.length - 1].id;
+      }
+      allReactingUsers = [...allReactingUsers, ...newReactingUsers];
+    }
+
+    return allReactingUsers.map((x) => ({
       channelId: channel,
       messageId: message,
+      emoji,
       authorId: x.id,
     }));
-    const hasNextPage = results.length === reactionsLimit;
-    const endCursor =
-      response.length > 0 ? response[response.length - 1].id : null;
-    const pageInfo = {hasNextPage, endCursor};
-    return {results, pageInfo};
+  }
+}
+
+function failIfMissing(response: Response) {
+  if (response.status === 404) {
+    throw new Error(`404 Not Found on: ${response.url}; maybe bad serverUrl?`);
+  }
+  if (response.status === 403) {
+    throw new Error(`403 Forbidden: bad API username or key?\n${response.url}`);
+  }
+  if (response.status === 410) {
+    throw new Error(`410 Gone`);
+  }
+}
+
+function failForNotOk(response: Response) {
+  if (!response.ok) {
+    throw new Error(`not OK status ${response.status} on ${response.url}`);
   }
 }
