+++ src/plugins/discord/plugin.js
