+++ src/plugins/discord/reactionWeights.js
