+++ src/plugins/discord/snapshots/aHR0cHM6Ly9kaXNjb3JkYXBwLmNvbS9hcGkvZ3VpbGRzLzY3ODM0ODk4MDYzOTQ5ODQyOC9jaGFubmVscw
@@ -21,7 +21,7 @@
   },
   {
     "id": "678348980849213472",
-    "last_message_id": "728740244131938314",
+    "last_message_id": "761706197875032104",
     "type": 0,
     "name": "general",
     "position": 0,
@@ -59,7 +59,7 @@
   },
   {
     "id": "678696874869522446",
-    "last_message_id": "678698977100824587",
+    "last_message_id": "761650961419993198",
     "type": 0,
     "name": "notmymessage",
     "position": 2,
