--- src/plugins/experimental-discord/mirror.test.js
@@ -1,28 +1,0 @@
-// @flow
-
-import Database from "better-sqlite3";
-import {snapshotFetcher} from "./mockSnapshotFetcher";
-import {SqliteMirrorRepository} from "./mirrorRepository";
-import {Mirror} from "./mirror";
-
-describe("plugins/experimental-discord/mirror", () => {
-  describe("smoke test", () => {
-    const guildId = "678348980639498428";
-    // const channelId = "678394406507905129";
-
-    it("should print", async () => {
-      // Given
-      const repo = new SqliteMirrorRepository(
-        new Database(":memory:"),
-        guildId
-      );
-      const api = snapshotFetcher();
-
-      // When
-      const mirror = new Mirror(repo, api, guildId);
-      await mirror.addMembers();
-      await mirror.addTextChannels();
-      // await mirror.addMessages(channelId, 10);
-    });
-  });
-});
